Lift the `Hero`, stepper wrappers, and submit button within `SetupFreeListings` to the upper component layers.
