+++ js/src/components/free-listings/setup-free-listings/form-content.js
@@ -1,67 +1,33 @@
 /**
- * External dependencies
- */
-import { __ } from '@wordpress/i18n';
-
-/**
  * Internal dependencies
  */
 import { useAdaptiveFormContext } from '~/components/adaptive-form';
-import StepContent from '~/components/stepper/step-content';
-import StepContentActions from '~/components/stepper/step-content-actions';
-import StepContentFooter from '~/components/stepper/step-content-footer';
 import ChooseAudienceSection from '~/components/free-listings/choose-audience-section';
 import ShippingRateSection from '~/components/shipping-rate-section';
 import ShippingTimeSection from '~/components/free-listings/configure-product-listings/shipping-time-section';
-import AppButton from '~/components/app-button';
 import OrderValueConditionSection from '~/components/order-value-condition-section';
 import isNonFreeShippingRate from '~/utils/isNonFreeShippingRate';
 
 /**
  * Form to configure free listigns.
- *
- * @param {Object} props React props.
- * @param {string} [props.submitLabel="Complete setup"] Submit button label.
  */
-const FormContent = ( {
-	submitLabel = __( 'Complete setup', 'google-listings-and-ads' ),
-} ) => {
-	const { values, isValidForm, handleSubmit, adapter } =
-		useAdaptiveFormContext();
+const FormContent = () => {
+	const { values } = useAdaptiveFormContext();
 
 	const shouldDisplayShippingTime = values.shipping_time === 'flat';
 	const shouldDisplayOrderValueCondition =
 		values.shipping_rate === 'flat' &&
 		values.shipping_country_rates.some( isNonFreeShippingRate );
 
-	const handleSubmitClick = ( event ) => {
-		if ( isValidForm ) {
-			return handleSubmit( event );
-		}
-
-		adapter.showValidation();
-	};
-
 	return (
-		<StepContent>
+		<>
 			<ChooseAudienceSection />
 			<ShippingRateSection />
 			{ shouldDisplayOrderValueCondition && (
 				<OrderValueConditionSection />
 			) }
 			{ shouldDisplayShippingTime && <ShippingTimeSection /> }
-			<StepContentFooter>
-				<StepContentActions>
-					<AppButton
-						isPrimary
-						loading={ adapter.isSubmitting }
-						onClick={ handleSubmitClick }
-					>
-						{ submitLabel }
-					</AppButton>
-				</StepContentActions>
-			</StepContentFooter>
-		</StepContent>
+		</>
 	);
 };
